linting

+++ src/sd.c
@@ -22,8 +22,8 @@
 #include "stm32mp13xx_hal_sd.h"
 #include "stm32mp13xx_ll_sdmmc.h"
 #include <inttypes.h>
-#include <stdint.h>
 #include <stdbool.h>
+#include <stdint.h>
 
 #define BLOCK_SIZE 512U
 #define DDR_SIZE   0x20000000U // 512 MB
@@ -96,14 +96,16 @@ void sd_read(uint32_t lba, uint32_t num_blocks, uint32_t dest_addr)
 
    // starting address sanity check
    if (dest_addr < DRAM_MEM_BASE) {
-      my_printf("Note: Adjusting DDR addr from 0x%" PRIX32 " to 0x%" PRIX32 "\r\n",
-            dest_addr, DRAM_MEM_BASE);
+      my_printf("Note: Adjusting DDR addr from 0x%" PRIX32 " to 0x%" PRIX32
+                "\r\n",
+                dest_addr, (uint32_t)DRAM_MEM_BASE);
       dest_addr = DRAM_MEM_BASE;
    }
 
    // check for overflow / underflow
-   const uint64_t end_addr = (uint64_t)dest_addr + (uint64_t)num_blocks * BLOCK_SIZE;
-   const bool overfl = end_addr > (uint64_t)DRAM_MEM_BASE + DDR_SIZE;
+   const uint64_t end_addr =
+       (uint64_t)dest_addr + ((uint64_t)num_blocks * BLOCK_SIZE);
+   const bool overfl  = end_addr > (uint64_t)DRAM_MEM_BASE + DDR_SIZE;
    const bool underfl = end_addr < DRAM_MEM_BASE;
    if (overfl || underfl) {
       my_printf("ERROR: SD read would overflow DDR!\r\n");
